requirements.txt
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
aniso8601==9.0.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.2.2
django-cors-headers==4.1.0
django-crontab==0.7.1
django-graphql-jwt==0.3.4
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
dotenv==0.0.5
frozenlist==1.3.3
graphene==3.2.2
graphene-django==3.1.2
graphql-core==3.2.3
graphql-relay==3.2.0
idna==3.4
multidict==6.0.4
mysqlclient==2.2.0
Pillow==9.5.0
promise==2.3
PyJWT==2.7.0
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
twilio==8.3.0
typing_extensions==4.6.3
urllib3==2.0.3
yarl==1.9.2
load-dotenv